Add the Morris screening method
This implements the OAT Morris method, a screening method which is based on
elementary effects to define sensitivity measures.

Contents:
 - MorrisExperiment class: it provides input designs for the sensitivity
   analysis. Both p-level grid (MorrisExperimentGrid) and centered LHS
   initial designs (MorrisExperimentLHS) are supported as starting argument
   for the definition of paths;
 - Morris class: it computes the elementary effects and its relevant risk
   indicators (mean, absolute mean, standard deviation);
 - A 'plot_sensitivity' python module, to plot elementary effects. It relies
   on matplotlib;
 - The MorrisFunction class, which inherits from OpenTURNSPythonFunction. It
   implements the Morris function (used for validation);
 - Full documentation: docstrings, use cases, examples, validation guide.

The split between MorrisExperiment and Morris classes helps when using an
external model.  In this case, one can call MorrisExperiment::generate method
to generate an input sample, evaluate it outside of this module, and pass
both input and output samples to Morris constructor.

Also, as OpenTURNS does not provide centered LHS, a static GenerateLHS method
is added in this module to define such designs.